tweaks

+++ lsx.c
@@ -3,7 +3,6 @@
 #include <limits.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
 #include <unistd.h>
 #include <sys/stat.h>
 
@@ -32,7 +31,7 @@ lsx(const char *dir) {
 		return;
 	}
 	while((d = readdir(dp)))
-		if(snprintf(buf, sizeof buf, "%s/%s", dir, d->d_name) < (ssize_t)sizeof buf
+		if(snprintf(buf, sizeof buf, "%s/%s", dir, d->d_name) < (int)sizeof buf
 		&& !stat(buf, &st) && S_ISREG(st.st_mode) && access(buf, X_OK) == 0)
 			puts(d->d_name);
 	closedir(dp);
